Refactored EventCard and renamed few things
