+++ back/src/database/associations.ts
@@ -17,11 +17,16 @@ const initAssociations = function () {
     Specie.hasMany(Animal, {foreignKey: 'specieId', sourceKey: 'id'});
     Animal.belongsTo(Specie, {foreignKey: 'specieId', targetKey: 'id'});
     Animal.belongsToMany(PetEvent, {as: 'events', through: 'eventAnimals', foreignKey: 'animalId'});
+
+
     PetEvent.belongsToMany(Animal, {as: 'attendees',through: 'eventAnimals', foreignKey: 'eventId'});
     Specie.belongsToMany(PetEvent, {as: 'events', through: 'eventSpecies', foreignKey: 'specieId'});
     PetEvent.belongsToMany(Specie, {as: 'authorizedSpecies', through: 'eventSpecies', foreignKey: 'eventId'});
     Animal.hasMany(AnimalPicture,{foreignKey: 'animalId', sourceKey: 'id'});
     AnimalPicture.belongsTo(Animal,{foreignKey: 'animalId', targetKey: 'id'});
+
+    User.hasMany(PetEvent, {foreignKey: 'userId', sourceKey: 'id', as:"events"});
+    PetEvent.belongsTo(User, {foreignKey: 'userId', targetKey: 'id'});
 
     User.hasMany(EventComment, {foreignKey: 'userId', sourceKey: 'id', as:"eventComments"});
     EventComment.belongsTo(User, {foreignKey: 'userId', targetKey: 'id'});
