+++ front/src/components/EventCard.tsx
@@ -1,7 +1,7 @@
 import React, {Component} from 'react';
 import peche from '../resources/events/peche.jpg';
 import { Link } from 'react-router-dom';
-import {PetEvent} from '../models/PetEvent';
+import {PetEvent} from '../models';
 
 interface EventCardProps {
   event:PetEvent;
@@ -19,10 +19,11 @@ export class EventCard extends Component<EventCardProps,{}> {
           <img src={peche} className="card-img-top" alt="Peche" />
           <div className="card-body">
             <h5 className="card-title">{event.name}</h5>
-            <h6 className="card-title">Le 25 mai</h6>
+            <h6 className="card-title">Début : {event.beginDate.toDateString()}</h6>
+            <h6 className="card-title">Fin : {event.endDate.toDateString()}</h6>
             <p className="card-text">{event.description}</p>
-            <h6 className="card-title">Organisé par {event.userId}</h6>
-            <Link to="/events/1" className="btn btn-primary">Détails</Link>
+            <h6 className="card-title">Organisé par {event.user ? event.user.username : "Inconnu(e)"}</h6>
+            <Link to={`/events/${event.id}`} className="btn btn-primary">Détails</Link>
           </div>
         </div>
       </div>
