+++ front/src/pages/index.ts
@@ -1,8 +1,8 @@
 export * from './HomePage';
 export * from './LoginPage';
 export * from './RegisterPage';
-export * from './EventList';
-export * from './EventDetails';
+export * from './EventsSearchPage';
+export * from './EventPage';
 export * from './ServicesPage';
 export * from './AnimalPage';
 export * from './NotFoundPage'
